// src/store/sidemenu/project_management/CompareQuotation/cq.types.ts
type UUID = string;
type DateString = string;
type DateTimeString = string;
// ...
interface Project {
	id: UUID;
	name: string | null;
	status: number;
	status_name: string;
	created_on: DateTimeString;
}

interface Item {
	name: string;
	id: string;
	value: string;
	label: string;
}

interface Unit {
	id: string;
	name: string;
	value: string;
	label: string;
}
interface Make {
	id: string;
	code: string;
	name: string;
	created_on: string;
	modified_on: string;
}

export interface CompareQuotationItem {
	id?: UUID;
	code: string;
	deliverydate: string;
	description: string;
	dodelete: boolean;
	comparequotation: CompareQuotation | string;
	make: Make;
	price: number | string | null;
	total_price: number | string | null;
	item: Item;
	unit: Unit;
	vendor: Vendor;
	qty: number | string;
	available_qty?: number;
	margin_value?: number;
	quotation?: Quotation;
	quotationitem?: QuotationItem;
	purchase_enquiry_item?: PQItem;
	originalqty: number | string;
	created_on?: DateTimeString;
	modified_on?: DateTimeString;
}

export interface CompareQuotation {
	id?: UUID;
	title?: string;
	code?: string;
	date?: DateString | null;
	project?: Project;
	currency?: Item;
	exchange_rate?: string;
	project_id?: UUID;
	purchase_enquiry?: miniType & PurchaseEnquiry;
	purchase_enquiry_id?: UUID;
	description?: string | null;
	total?: number | string | null;
	deliverydate?: DateString;
	vendor?: miniType & Vendor;
	vendor_id?: UUID;
	quotation?: Quotation;
	quotationitem?: PQItem;
	quotation_id?: UUID;
	comparequotationitems?: CompareQuotationItem[];
	cqstatus?: number;
	cqstatus_name?: string;
	qty?: string | number;
	created_on?: DateTimeString;
	created_by?: {
		fullname: string;
	};
	modified_on?: DateTimeString;
	purchase_enquiry_item?:
		| {
				id: string;
		  }
		| string;
	purchase_enquiry_item_id?: string;

	quotationitems?: {
		id?: string;
		dodelete?: boolean;
		price?: string;
		total_price?: string;
		item?: miniType;
		qty?: string;
	}[];

	authorized_status_name?: string;
	authorized_status?: number;
	current_authorized_level?: number;
}

interface Minirendertype {
	label: string;
	value: string;
}

export interface CheckedListItem {
	id?: string;
	temp_q: Quotation;
	quotationitem: CompareQuotationItem;
	vendor: Vendor;
	purchase_enquiry_item_id: string;
	dodelete: boolean;
}

export type VendorRelatedItem = Vendor & {
	vendorItems?: CompareQuotationItem[];
	deliverydate?: string;
	total?: number;
	quotation?: Quotation;
};

export interface CompareQuotationInitialState extends BaseInitialState {
	compareQuotationList: CompareQuotation[];
	compareQuotationCount: number;
	isGeneratePoModalOpen: boolean;
	selectedVendor?: VendorRelatedItem | null;
	pageParams: PageParamsTypes & {
		start_date?: string;
		end_date?: string;
		vendor?: string;
		project_id: string | number;
		purchase_enquiry?: Minirendertype | null;
		quotation?: Minirendertype | null;
		cqstatus?: Minirendertype | null;
	};
	POByPE: Order[];
	vendorRelatedItems?: VendorRelatedItem[];
	checkedList?: any;
	selectedData: CompareQuotation;
	PEaganistCQ: {
		list: Array<any>;
		count: number;
		loading: boolean;
		miniParams: PageParamsTypes;
	};
	isFilterOpen: boolean;

	// PQ
	PQLoading: boolean;

	purchaseQuotationList: Quotation[];
	purchaseQuotationCount: number;
	PQpageParams: PageParamsTypes & {
		start_date?: string;
		end_date?: string;
		purchase_enquiry?: string | null;
		vendor?: miniType | null;
		project_id?: string | number;
		qstatus?: miniType | null;
	};

	// PE

	peSelectedData: PurchaseEnquiry;
	peLoading: boolean;
}